Support for other than Ensembl seqlevel style

The methods and functions on this help page allow to integrate EnsDb objects and the annotations they provide with other Bioconductor annotation packages that base on chromosome names (seqlevels) that are different from those defined by Ensembl.

# S4 method for EnsDb
seqlevelsStyle(x)

# S4 method for EnsDb
seqlevelsStyle(x) <- value

# S4 method for EnsDb
supportedSeqlevelsStyles(x)

Arguments

(In alphabetic order)

value

For seqlevelsStyle<-: a character string specifying the seqlevels style that should be set. Use the supportedSeqlevelsStyle to list all available and supported seqlevel styles.

x

An EnsDb instance.

Methods and Functions

seqlevelsStyle

Get the style of the seqlevels in which results returned from the EnsDb object are encoded. By default, and internally, seqnames as provided by Ensembl are used.

The method returns a character string specifying the currently used seqlevelstyle.

seqlevelsStyle<-

Change the style of the seqlevels in which results returned from the EnsDb object are encoded. Changing the seqlevels helps integrating annotations from EnsDb objects e.g. with annotations from packages that base on UCSC annotations.

supportedSeqlevelsStyles

Lists all seqlevel styles for which mappings between seqlevel styles are available in the GenomeInfoDb package.

The method returns a character vector with supported seqlevel styles for the organism of the EnsDb object.

Note

The mapping between different seqname styles is performed based on data provided by the GenomeInfoDb package. Note that in most instances no mapping is provided for seqnames other than for primary chromosomes. By default functions from the ensembldb package return the original seqname is in such cases. This behaviour can be changed with the ensembldb.seqnameNotFound global option. For the special keyword "ORIGINAL" (the default), the original seqnames are returned, for "MISSING" an error is thrown if a seqname can not be mapped. In all other cases, the value of the option is returned as seqname if no mapping is available (e.g. setting options(ensembldb.seqnameNotFound=NA) returns an NA if the seqname is not mappable).

Value

For seqlevelsStyle: see method description above. For supportedSeqlevelsStyles: see method description above.

Author

Johannes Rainer

See also

Examples

library(EnsDb.Hsapiens.v86)
edb <- EnsDb.Hsapiens.v86

## Get the internal, default seqlevel style.
seqlevelsStyle(edb)
#> [1] "Ensembl"

## Get the seqlevels from the database.
seqlevels(edb)
#>   [1] "1"                                     
#>   [2] "10"                                    
#>   [3] "11"                                    
#>   [4] "12"                                    
#>   [5] "13"                                    
#>   [6] "14"                                    
#>   [7] "15"                                    
#>   [8] "16"                                    
#>   [9] "17"                                    
#>  [10] "18"                                    
#>  [11] "19"                                    
#>  [12] "2"                                     
#>  [13] "20"                                    
#>  [14] "21"                                    
#>  [15] "22"                                    
#>  [16] "3"                                     
#>  [17] "4"                                     
#>  [18] "5"                                     
#>  [19] "6"                                     
#>  [20] "7"                                     
#>  [21] "8"                                     
#>  [22] "9"                                     
#>  [23] "CHR_HG107_PATCH"                       
#>  [24] "CHR_HG126_PATCH"                       
#>  [25] "CHR_HG1311_PATCH"                      
#>  [26] "CHR_HG1342_HG2282_PATCH"               
#>  [27] "CHR_HG1362_PATCH"                      
#>  [28] "CHR_HG142_HG150_NOVEL_TEST"            
#>  [29] "CHR_HG151_NOVEL_TEST"                  
#>  [30] "CHR_HG1651_PATCH"                      
#>  [31] "CHR_HG1832_PATCH"                      
#>  [32] "CHR_HG2021_PATCH"                      
#>  [33] "CHR_HG2022_PATCH"                      
#>  [34] "CHR_HG2023_PATCH"                      
#>  [35] "CHR_HG2030_PATCH"                      
#>  [36] "CHR_HG2058_PATCH"                      
#>  [37] "CHR_HG2062_PATCH"                      
#>  [38] "CHR_HG2063_PATCH"                      
#>  [39] "CHR_HG2066_PATCH"                      
#>  [40] "CHR_HG2072_PATCH"                      
#>  [41] "CHR_HG2095_PATCH"                      
#>  [42] "CHR_HG2104_PATCH"                      
#>  [43] "CHR_HG2116_PATCH"                      
#>  [44] "CHR_HG2128_PATCH"                      
#>  [45] "CHR_HG2191_PATCH"                      
#>  [46] "CHR_HG2213_PATCH"                      
#>  [47] "CHR_HG2217_PATCH"                      
#>  [48] "CHR_HG2232_PATCH"                      
#>  [49] "CHR_HG2233_PATCH"                      
#>  [50] "CHR_HG2235_PATCH"                      
#>  [51] "CHR_HG2239_PATCH"                      
#>  [52] "CHR_HG2247_PATCH"                      
#>  [53] "CHR_HG2249_PATCH"                      
#>  [54] "CHR_HG2288_HG2289_PATCH"               
#>  [55] "CHR_HG2290_PATCH"                      
#>  [56] "CHR_HG2291_PATCH"                      
#>  [57] "CHR_HG2334_PATCH"                      
#>  [58] "CHR_HG26_PATCH"                        
#>  [59] "CHR_HG986_PATCH"                       
#>  [60] "CHR_HSCHR10_1_CTG1"                    
#>  [61] "CHR_HSCHR10_1_CTG2"                    
#>  [62] "CHR_HSCHR10_1_CTG3"                    
#>  [63] "CHR_HSCHR10_1_CTG4"                    
#>  [64] "CHR_HSCHR10_1_CTG6"                    
#>  [65] "CHR_HSCHR11_1_CTG1_2"                  
#>  [66] "CHR_HSCHR11_1_CTG5"                    
#>  [67] "CHR_HSCHR11_1_CTG6"                    
#>  [68] "CHR_HSCHR11_1_CTG7"                    
#>  [69] "CHR_HSCHR11_1_CTG8"                    
#>  [70] "CHR_HSCHR11_2_CTG1"                    
#>  [71] "CHR_HSCHR11_2_CTG1_1"                  
#>  [72] "CHR_HSCHR11_3_CTG1"                    
#>  [73] "CHR_HSCHR12_1_CTG1"                    
#>  [74] "CHR_HSCHR12_1_CTG2_1"                  
#>  [75] "CHR_HSCHR12_2_CTG1"                    
#>  [76] "CHR_HSCHR12_2_CTG2"                    
#>  [77] "CHR_HSCHR12_2_CTG2_1"                  
#>  [78] "CHR_HSCHR12_3_CTG2"                    
#>  [79] "CHR_HSCHR12_3_CTG2_1"                  
#>  [80] "CHR_HSCHR12_4_CTG2"                    
#>  [81] "CHR_HSCHR12_4_CTG2_1"                  
#>  [82] "CHR_HSCHR12_5_CTG2"                    
#>  [83] "CHR_HSCHR12_5_CTG2_1"                  
#>  [84] "CHR_HSCHR12_6_CTG2_1"                  
#>  [85] "CHR_HSCHR13_1_CTG1"                    
#>  [86] "CHR_HSCHR13_1_CTG3"                    
#>  [87] "CHR_HSCHR13_1_CTG5"                    
#>  [88] "CHR_HSCHR13_1_CTG8"                    
#>  [89] "CHR_HSCHR14_1_CTG1"                    
#>  [90] "CHR_HSCHR14_2_CTG1"                    
#>  [91] "CHR_HSCHR14_3_CTG1"                    
#>  [92] "CHR_HSCHR14_7_CTG1"                    
#>  [93] "CHR_HSCHR15_1_CTG1"                    
#>  [94] "CHR_HSCHR15_1_CTG3"                    
#>  [95] "CHR_HSCHR15_1_CTG8"                    
#>  [96] "CHR_HSCHR15_2_CTG3"                    
#>  [97] "CHR_HSCHR15_2_CTG8"                    
#>  [98] "CHR_HSCHR15_3_CTG3"                    
#>  [99] "CHR_HSCHR15_3_CTG8"                    
#> [100] "CHR_HSCHR15_4_CTG8"                    
#> [101] "CHR_HSCHR15_5_CTG8"                    
#> [102] "CHR_HSCHR15_6_CTG8"                    
#> [103] "CHR_HSCHR16_1_CTG1"                    
#> [104] "CHR_HSCHR16_1_CTG3_1"                  
#> [105] "CHR_HSCHR16_2_CTG3_1"                  
#> [106] "CHR_HSCHR16_3_CTG1"                    
#> [107] "CHR_HSCHR16_4_CTG1"                    
#> [108] "CHR_HSCHR16_4_CTG3_1"                  
#> [109] "CHR_HSCHR16_5_CTG1"                    
#> [110] "CHR_HSCHR16_CTG2"                      
#> [111] "CHR_HSCHR17_10_CTG4"                   
#> [112] "CHR_HSCHR17_1_CTG1"                    
#> [113] "CHR_HSCHR17_1_CTG2"                    
#> [114] "CHR_HSCHR17_1_CTG4"                    
#> [115] "CHR_HSCHR17_1_CTG5"                    
#> [116] "CHR_HSCHR17_1_CTG9"                    
#> [117] "CHR_HSCHR17_2_CTG1"                    
#> [118] "CHR_HSCHR17_2_CTG2"                    
#> [119] "CHR_HSCHR17_2_CTG4"                    
#> [120] "CHR_HSCHR17_2_CTG5"                    
#> [121] "CHR_HSCHR17_3_CTG2"                    
#> [122] "CHR_HSCHR17_3_CTG4"                    
#> [123] "CHR_HSCHR17_4_CTG4"                    
#> [124] "CHR_HSCHR17_5_CTG4"                    
#> [125] "CHR_HSCHR17_6_CTG4"                    
#> [126] "CHR_HSCHR17_7_CTG4"                    
#> [127] "CHR_HSCHR17_8_CTG4"                    
#> [128] "CHR_HSCHR17_9_CTG4"                    
#> [129] "CHR_HSCHR18_1_CTG1_1"                  
#> [130] "CHR_HSCHR18_1_CTG2_1"                  
#> [131] "CHR_HSCHR18_2_CTG1_1"                  
#> [132] "CHR_HSCHR18_2_CTG2"                    
#> [133] "CHR_HSCHR18_2_CTG2_1"                  
#> [134] "CHR_HSCHR18_3_CTG2_1"                  
#> [135] "CHR_HSCHR18_5_CTG1_1"                  
#> [136] "CHR_HSCHR18_ALT21_CTG2_1"              
#> [137] "CHR_HSCHR18_ALT2_CTG2_1"               
#> [138] "CHR_HSCHR19KIR_ABC08_A1_HAP_CTG3_1"    
#> [139] "CHR_HSCHR19KIR_ABC08_AB_HAP_C_P_CTG3_1"
#> [140] "CHR_HSCHR19KIR_ABC08_AB_HAP_T_P_CTG3_1"
#> [141] "CHR_HSCHR19KIR_FH05_A_HAP_CTG3_1"      
#> [142] "CHR_HSCHR19KIR_FH05_B_HAP_CTG3_1"      
#> [143] "CHR_HSCHR19KIR_FH06_A_HAP_CTG3_1"      
#> [144] "CHR_HSCHR19KIR_FH06_BA1_HAP_CTG3_1"    
#> [145] "CHR_HSCHR19KIR_FH08_A_HAP_CTG3_1"      
#> [146] "CHR_HSCHR19KIR_FH08_BAX_HAP_CTG3_1"    
#> [147] "CHR_HSCHR19KIR_FH13_A_HAP_CTG3_1"      
#> [148] "CHR_HSCHR19KIR_FH13_BA2_HAP_CTG3_1"    
#> [149] "CHR_HSCHR19KIR_FH15_A_HAP_CTG3_1"      
#> [150] "CHR_HSCHR19KIR_FH15_B_HAP_CTG3_1"      
#> [151] "CHR_HSCHR19KIR_G085_A_HAP_CTG3_1"      
#> [152] "CHR_HSCHR19KIR_G085_BA1_HAP_CTG3_1"    
#> [153] "CHR_HSCHR19KIR_G248_A_HAP_CTG3_1"      
#> [154] "CHR_HSCHR19KIR_G248_BA2_HAP_CTG3_1"    
#> [155] "CHR_HSCHR19KIR_GRC212_AB_HAP_CTG3_1"   
#> [156] "CHR_HSCHR19KIR_GRC212_BA1_HAP_CTG3_1"  
#> [157] "CHR_HSCHR19KIR_LUCE_A_HAP_CTG3_1"      
#> [158] "CHR_HSCHR19KIR_LUCE_BDEL_HAP_CTG3_1"   
#> [159] "CHR_HSCHR19KIR_RP5_B_HAP_CTG3_1"       
#> [160] "CHR_HSCHR19KIR_RSH_A_HAP_CTG3_1"       
#> [161] "CHR_HSCHR19KIR_RSH_BA2_HAP_CTG3_1"     
#> [162] "CHR_HSCHR19KIR_T7526_A_HAP_CTG3_1"     
#> [163] "CHR_HSCHR19KIR_T7526_BDEL_HAP_CTG3_1"  
#> [164] "CHR_HSCHR19LRC_COX1_CTG3_1"            
#> [165] "CHR_HSCHR19LRC_COX2_CTG3_1"            
#> [166] "CHR_HSCHR19LRC_LRC_I_CTG3_1"           
#> [167] "CHR_HSCHR19LRC_LRC_J_CTG3_1"           
#> [168] "CHR_HSCHR19LRC_LRC_S_CTG3_1"           
#> [169] "CHR_HSCHR19LRC_LRC_T_CTG3_1"           
#> [170] "CHR_HSCHR19LRC_PGF1_CTG3_1"            
#> [171] "CHR_HSCHR19LRC_PGF2_CTG3_1"            
#> [172] "CHR_HSCHR19_1_CTG2"                    
#> [173] "CHR_HSCHR19_1_CTG3_1"                  
#> [174] "CHR_HSCHR19_2_CTG2"                    
#> [175] "CHR_HSCHR19_2_CTG3_1"                  
#> [176] "CHR_HSCHR19_3_CTG2"                    
#> [177] "CHR_HSCHR19_3_CTG3_1"                  
#> [178] "CHR_HSCHR19_4_CTG2"                    
#> [179] "CHR_HSCHR19_4_CTG3_1"                  
#> [180] "CHR_HSCHR19_5_CTG2"                    
#> [181] "CHR_HSCHR1_1_CTG11"                    
#> [182] "CHR_HSCHR1_1_CTG3"                     
#> [183] "CHR_HSCHR1_1_CTG31"                    
#> [184] "CHR_HSCHR1_1_CTG32_1"                  
#> [185] "CHR_HSCHR1_2_CTG3"                     
#> [186] "CHR_HSCHR1_2_CTG31"                    
#> [187] "CHR_HSCHR1_2_CTG32_1"                  
#> [188] "CHR_HSCHR1_3_CTG3"                     
#> [189] "CHR_HSCHR1_3_CTG31"                    
#> [190] "CHR_HSCHR1_3_CTG32_1"                  
#> [191] "CHR_HSCHR1_4_CTG3"                     
#> [192] "CHR_HSCHR1_4_CTG31"                    
#> [193] "CHR_HSCHR1_5_CTG3"                     
#> [194] "CHR_HSCHR1_5_CTG32_1"                  
#> [195] "CHR_HSCHR1_ALT2_1_CTG32_1"             
#> [196] "CHR_HSCHR20_1_CTG1"                    
#> [197] "CHR_HSCHR20_1_CTG2"                    
#> [198] "CHR_HSCHR20_1_CTG3"                    
#> [199] "CHR_HSCHR20_1_CTG4"                    
#> [200] "CHR_HSCHR21_2_CTG1_1"                  
#> [201] "CHR_HSCHR21_3_CTG1_1"                  
#> [202] "CHR_HSCHR21_4_CTG1_1"                  
#> [203] "CHR_HSCHR21_5_CTG2"                    
#> [204] "CHR_HSCHR21_6_CTG1_1"                  
#> [205] "CHR_HSCHR21_8_CTG1_1"                  
#> [206] "CHR_HSCHR22_1_CTG1"                    
#> [207] "CHR_HSCHR22_1_CTG2"                    
#> [208] "CHR_HSCHR22_1_CTG3"                    
#> [209] "CHR_HSCHR22_1_CTG4"                    
#> [210] "CHR_HSCHR22_1_CTG5"                    
#> [211] "CHR_HSCHR22_1_CTG6"                    
#> [212] "CHR_HSCHR22_1_CTG7"                    
#> [213] "CHR_HSCHR22_2_CTG1"                    
#> [214] "CHR_HSCHR22_3_CTG1"                    
#> [215] "CHR_HSCHR22_4_CTG1"                    
#> [216] "CHR_HSCHR22_5_CTG1"                    
#> [217] "CHR_HSCHR22_6_CTG1"                    
#> [218] "CHR_HSCHR22_7_CTG1"                    
#> [219] "CHR_HSCHR22_8_CTG1"                    
#> [220] "CHR_HSCHR2_1_CTG1"                     
#> [221] "CHR_HSCHR2_1_CTG15"                    
#> [222] "CHR_HSCHR2_1_CTG5"                     
#> [223] "CHR_HSCHR2_1_CTG7"                     
#> [224] "CHR_HSCHR2_1_CTG7_2"                   
#> [225] "CHR_HSCHR2_2_CTG1"                     
#> [226] "CHR_HSCHR2_2_CTG15"                    
#> [227] "CHR_HSCHR2_2_CTG7"                     
#> [228] "CHR_HSCHR2_2_CTG7_2"                   
#> [229] "CHR_HSCHR2_3_CTG1"                     
#> [230] "CHR_HSCHR2_3_CTG15"                    
#> [231] "CHR_HSCHR2_3_CTG7_2"                   
#> [232] "CHR_HSCHR2_4_CTG1"                     
#> [233] "CHR_HSCHR2_6_CTG7_2"                   
#> [234] "CHR_HSCHR3_1_CTG1"                     
#> [235] "CHR_HSCHR3_1_CTG2_1"                   
#> [236] "CHR_HSCHR3_1_CTG3"                     
#> [237] "CHR_HSCHR3_2_CTG2_1"                   
#> [238] "CHR_HSCHR3_2_CTG3"                     
#> [239] "CHR_HSCHR3_3_CTG1"                     
#> [240] "CHR_HSCHR3_3_CTG3"                     
#> [241] "CHR_HSCHR3_4_CTG2_1"                   
#> [242] "CHR_HSCHR3_4_CTG3"                     
#> [243] "CHR_HSCHR3_5_CTG2_1"                   
#> [244] "CHR_HSCHR3_5_CTG3"                     
#> [245] "CHR_HSCHR3_6_CTG3"                     
#> [246] "CHR_HSCHR3_7_CTG3"                     
#> [247] "CHR_HSCHR3_8_CTG3"                     
#> [248] "CHR_HSCHR3_9_CTG3"                     
#> [249] "CHR_HSCHR4_11_CTG12"                   
#> [250] "CHR_HSCHR4_1_CTG12"                    
#> [251] "CHR_HSCHR4_1_CTG4"                     
#> [252] "CHR_HSCHR4_1_CTG6"                     
#> [253] "CHR_HSCHR4_1_CTG9"                     
#> [254] "CHR_HSCHR4_2_CTG12"                    
#> [255] "CHR_HSCHR4_2_CTG4"                     
#> [256] "CHR_HSCHR4_3_CTG12"                    
#> [257] "CHR_HSCHR4_4_CTG12"                    
#> [258] "CHR_HSCHR4_5_CTG12"                    
#> [259] "CHR_HSCHR4_6_CTG12"                    
#> [260] "CHR_HSCHR4_7_CTG12"                    
#> [261] "CHR_HSCHR4_8_CTG12"                    
#> [262] "CHR_HSCHR4_9_CTG12"                    
#> [263] "CHR_HSCHR5_1_CTG1"                     
#> [264] "CHR_HSCHR5_1_CTG1_1"                   
#> [265] "CHR_HSCHR5_1_CTG5"                     
#> [266] "CHR_HSCHR5_2_CTG1"                     
#> [267] "CHR_HSCHR5_2_CTG1_1"                   
#> [268] "CHR_HSCHR5_2_CTG5"                     
#> [269] "CHR_HSCHR5_3_CTG1"                     
#> [270] "CHR_HSCHR5_3_CTG5"                     
#> [271] "CHR_HSCHR5_4_CTG1"                     
#> [272] "CHR_HSCHR5_4_CTG1_1"                   
#> [273] "CHR_HSCHR5_5_CTG1"                     
#> [274] "CHR_HSCHR5_6_CTG1"                     
#> [275] "CHR_HSCHR5_7_CTG1"                     
#> [276] "CHR_HSCHR6_1_CTG10"                    
#> [277] "CHR_HSCHR6_1_CTG2"                     
#> [278] "CHR_HSCHR6_1_CTG3"                     
#> [279] "CHR_HSCHR6_1_CTG4"                     
#> [280] "CHR_HSCHR6_1_CTG5"                     
#> [281] "CHR_HSCHR6_1_CTG6"                     
#> [282] "CHR_HSCHR6_1_CTG7"                     
#> [283] "CHR_HSCHR6_1_CTG8"                     
#> [284] "CHR_HSCHR6_1_CTG9"                     
#> [285] "CHR_HSCHR6_8_CTG1"                     
#> [286] "CHR_HSCHR6_MHC_APD_CTG1"               
#> [287] "CHR_HSCHR6_MHC_COX_CTG1"               
#> [288] "CHR_HSCHR6_MHC_DBB_CTG1"               
#> [289] "CHR_HSCHR6_MHC_MANN_CTG1"              
#> [290] "CHR_HSCHR6_MHC_MCF_CTG1"               
#> [291] "CHR_HSCHR6_MHC_QBL_CTG1"               
#> [292] "CHR_HSCHR6_MHC_SSTO_CTG1"              
#> [293] "CHR_HSCHR7_1_CTG1"                     
#> [294] "CHR_HSCHR7_1_CTG4_4"                   
#> [295] "CHR_HSCHR7_1_CTG6"                     
#> [296] "CHR_HSCHR7_1_CTG7"                     
#> [297] "CHR_HSCHR7_2_CTG1"                     
#> [298] "CHR_HSCHR7_2_CTG4_4"                   
#> [299] "CHR_HSCHR7_2_CTG6"                     
#> [300] "CHR_HSCHR7_2_CTG7"                     
#> [301] "CHR_HSCHR7_3_CTG6"                     
#> [302] "CHR_HSCHR8_1_CTG1"                     
#> [303] "CHR_HSCHR8_1_CTG6"                     
#> [304] "CHR_HSCHR8_1_CTG7"                     
#> [305] "CHR_HSCHR8_2_CTG1"                     
#> [306] "CHR_HSCHR8_2_CTG7"                     
#> [307] "CHR_HSCHR8_3_CTG1"                     
#> [308] "CHR_HSCHR8_3_CTG7"                     
#> [309] "CHR_HSCHR8_4_CTG1"                     
#> [310] "CHR_HSCHR8_4_CTG7"                     
#> [311] "CHR_HSCHR8_5_CTG1"                     
#> [312] "CHR_HSCHR8_5_CTG7"                     
#> [313] "CHR_HSCHR8_6_CTG1"                     
#> [314] "CHR_HSCHR8_7_CTG1"                     
#> [315] "CHR_HSCHR8_8_CTG1"                     
#> [316] "CHR_HSCHR8_9_CTG1"                     
#> [317] "CHR_HSCHR9_1_CTG1"                     
#> [318] "CHR_HSCHR9_1_CTG2"                     
#> [319] "CHR_HSCHR9_1_CTG3"                     
#> [320] "CHR_HSCHR9_1_CTG4"                     
#> [321] "CHR_HSCHR9_1_CTG5"                     
#> [322] "CHR_HSCHR9_1_CTG6"                     
#> [323] "CHR_HSCHRX_1_CTG3"                     
#> [324] "CHR_HSCHRX_2_CTG12"                    
#> [325] "CHR_HSCHRX_2_CTG3"                     
#> [326] "GL000009.2"                            
#> [327] "GL000194.1"                            
#> [328] "GL000195.1"                            
#> [329] "GL000205.2"                            
#> [330] "GL000213.1"                            
#> [331] "GL000216.2"                            
#> [332] "GL000218.1"                            
#> [333] "GL000219.1"                            
#> [334] "GL000220.1"                            
#> [335] "GL000225.1"                            
#> [336] "KI270442.1"                            
#> [337] "KI270711.1"                            
#> [338] "KI270713.1"                            
#> [339] "KI270721.1"                            
#> [340] "KI270726.1"                            
#> [341] "KI270727.1"                            
#> [342] "KI270728.1"                            
#> [343] "KI270731.1"                            
#> [344] "KI270733.1"                            
#> [345] "KI270734.1"                            
#> [346] "KI270744.1"                            
#> [347] "KI270750.1"                            
#> [348] "LRG_183"                               
#> [349] "LRG_187"                               
#> [350] "LRG_239"                               
#> [351] "LRG_311"                               
#> [352] "LRG_721"                               
#> [353] "LRG_741"                               
#> [354] "LRG_93"                                
#> [355] "MT"                                    
#> [356] "X"                                     
#> [357] "Y"                                     

## Get all supported mappings for the organism of the EnsDb.
supportedSeqlevelsStyles(edb)
#> NULL

## Change the seqlevels to UCSC style.
seqlevelsStyle(edb) <- "UCSC"
seqlevels(edb)
#> Warning: More than 5 seqnames with seqlevels style of the database (Ensembl) could not be mapped to the seqlevels style: UCSC!) Returning the orginal seqnames for these.
#>   [1] "chr1"                                  
#>   [2] "chr10"                                 
#>   [3] "chr11"                                 
#>   [4] "chr12"                                 
#>   [5] "chr13"                                 
#>   [6] "chr14"                                 
#>   [7] "chr15"                                 
#>   [8] "chr16"                                 
#>   [9] "chr17"                                 
#>  [10] "chr18"                                 
#>  [11] "chr19"                                 
#>  [12] "chr2"                                  
#>  [13] "chr20"                                 
#>  [14] "chr21"                                 
#>  [15] "chr22"                                 
#>  [16] "chr3"                                  
#>  [17] "chr4"                                  
#>  [18] "chr5"                                  
#>  [19] "chr6"                                  
#>  [20] "chr7"                                  
#>  [21] "chr8"                                  
#>  [22] "chr9"                                  
#>  [23] "CHR_HG107_PATCH"                       
#>  [24] "CHR_HG126_PATCH"                       
#>  [25] "CHR_HG1311_PATCH"                      
#>  [26] "CHR_HG1342_HG2282_PATCH"               
#>  [27] "CHR_HG1362_PATCH"                      
#>  [28] "CHR_HG142_HG150_NOVEL_TEST"            
#>  [29] "CHR_HG151_NOVEL_TEST"                  
#>  [30] "CHR_HG1651_PATCH"                      
#>  [31] "CHR_HG1832_PATCH"                      
#>  [32] "CHR_HG2021_PATCH"                      
#>  [33] "CHR_HG2022_PATCH"                      
#>  [34] "CHR_HG2023_PATCH"                      
#>  [35] "CHR_HG2030_PATCH"                      
#>  [36] "CHR_HG2058_PATCH"                      
#>  [37] "CHR_HG2062_PATCH"                      
#>  [38] "CHR_HG2063_PATCH"                      
#>  [39] "CHR_HG2066_PATCH"                      
#>  [40] "CHR_HG2072_PATCH"                      
#>  [41] "CHR_HG2095_PATCH"                      
#>  [42] "CHR_HG2104_PATCH"                      
#>  [43] "CHR_HG2116_PATCH"                      
#>  [44] "CHR_HG2128_PATCH"                      
#>  [45] "CHR_HG2191_PATCH"                      
#>  [46] "CHR_HG2213_PATCH"                      
#>  [47] "CHR_HG2217_PATCH"                      
#>  [48] "CHR_HG2232_PATCH"                      
#>  [49] "CHR_HG2233_PATCH"                      
#>  [50] "CHR_HG2235_PATCH"                      
#>  [51] "CHR_HG2239_PATCH"                      
#>  [52] "CHR_HG2247_PATCH"                      
#>  [53] "CHR_HG2249_PATCH"                      
#>  [54] "CHR_HG2288_HG2289_PATCH"               
#>  [55] "CHR_HG2290_PATCH"                      
#>  [56] "CHR_HG2291_PATCH"                      
#>  [57] "CHR_HG2334_PATCH"                      
#>  [58] "CHR_HG26_PATCH"                        
#>  [59] "CHR_HG986_PATCH"                       
#>  [60] "CHR_HSCHR10_1_CTG1"                    
#>  [61] "CHR_HSCHR10_1_CTG2"                    
#>  [62] "CHR_HSCHR10_1_CTG3"                    
#>  [63] "CHR_HSCHR10_1_CTG4"                    
#>  [64] "CHR_HSCHR10_1_CTG6"                    
#>  [65] "CHR_HSCHR11_1_CTG1_2"                  
#>  [66] "CHR_HSCHR11_1_CTG5"                    
#>  [67] "CHR_HSCHR11_1_CTG6"                    
#>  [68] "CHR_HSCHR11_1_CTG7"                    
#>  [69] "CHR_HSCHR11_1_CTG8"                    
#>  [70] "CHR_HSCHR11_2_CTG1"                    
#>  [71] "CHR_HSCHR11_2_CTG1_1"                  
#>  [72] "CHR_HSCHR11_3_CTG1"                    
#>  [73] "CHR_HSCHR12_1_CTG1"                    
#>  [74] "CHR_HSCHR12_1_CTG2_1"                  
#>  [75] "CHR_HSCHR12_2_CTG1"                    
#>  [76] "CHR_HSCHR12_2_CTG2"                    
#>  [77] "CHR_HSCHR12_2_CTG2_1"                  
#>  [78] "CHR_HSCHR12_3_CTG2"                    
#>  [79] "CHR_HSCHR12_3_CTG2_1"                  
#>  [80] "CHR_HSCHR12_4_CTG2"                    
#>  [81] "CHR_HSCHR12_4_CTG2_1"                  
#>  [82] "CHR_HSCHR12_5_CTG2"                    
#>  [83] "CHR_HSCHR12_5_CTG2_1"                  
#>  [84] "CHR_HSCHR12_6_CTG2_1"                  
#>  [85] "CHR_HSCHR13_1_CTG1"                    
#>  [86] "CHR_HSCHR13_1_CTG3"                    
#>  [87] "CHR_HSCHR13_1_CTG5"                    
#>  [88] "CHR_HSCHR13_1_CTG8"                    
#>  [89] "CHR_HSCHR14_1_CTG1"                    
#>  [90] "CHR_HSCHR14_2_CTG1"                    
#>  [91] "CHR_HSCHR14_3_CTG1"                    
#>  [92] "CHR_HSCHR14_7_CTG1"                    
#>  [93] "CHR_HSCHR15_1_CTG1"                    
#>  [94] "CHR_HSCHR15_1_CTG3"                    
#>  [95] "CHR_HSCHR15_1_CTG8"                    
#>  [96] "CHR_HSCHR15_2_CTG3"                    
#>  [97] "CHR_HSCHR15_2_CTG8"                    
#>  [98] "CHR_HSCHR15_3_CTG3"                    
#>  [99] "CHR_HSCHR15_3_CTG8"                    
#> [100] "CHR_HSCHR15_4_CTG8"                    
#> [101] "CHR_HSCHR15_5_CTG8"                    
#> [102] "CHR_HSCHR15_6_CTG8"                    
#> [103] "CHR_HSCHR16_1_CTG1"                    
#> [104] "CHR_HSCHR16_1_CTG3_1"                  
#> [105] "CHR_HSCHR16_2_CTG3_1"                  
#> [106] "CHR_HSCHR16_3_CTG1"                    
#> [107] "CHR_HSCHR16_4_CTG1"                    
#> [108] "CHR_HSCHR16_4_CTG3_1"                  
#> [109] "CHR_HSCHR16_5_CTG1"                    
#> [110] "CHR_HSCHR16_CTG2"                      
#> [111] "CHR_HSCHR17_10_CTG4"                   
#> [112] "CHR_HSCHR17_1_CTG1"                    
#> [113] "CHR_HSCHR17_1_CTG2"                    
#> [114] "CHR_HSCHR17_1_CTG4"                    
#> [115] "CHR_HSCHR17_1_CTG5"                    
#> [116] "CHR_HSCHR17_1_CTG9"                    
#> [117] "CHR_HSCHR17_2_CTG1"                    
#> [118] "CHR_HSCHR17_2_CTG2"                    
#> [119] "CHR_HSCHR17_2_CTG4"                    
#> [120] "CHR_HSCHR17_2_CTG5"                    
#> [121] "CHR_HSCHR17_3_CTG2"                    
#> [122] "CHR_HSCHR17_3_CTG4"                    
#> [123] "CHR_HSCHR17_4_CTG4"                    
#> [124] "CHR_HSCHR17_5_CTG4"                    
#> [125] "CHR_HSCHR17_6_CTG4"                    
#> [126] "CHR_HSCHR17_7_CTG4"                    
#> [127] "CHR_HSCHR17_8_CTG4"                    
#> [128] "CHR_HSCHR17_9_CTG4"                    
#> [129] "CHR_HSCHR18_1_CTG1_1"                  
#> [130] "CHR_HSCHR18_1_CTG2_1"                  
#> [131] "CHR_HSCHR18_2_CTG1_1"                  
#> [132] "CHR_HSCHR18_2_CTG2"                    
#> [133] "CHR_HSCHR18_2_CTG2_1"                  
#> [134] "CHR_HSCHR18_3_CTG2_1"                  
#> [135] "CHR_HSCHR18_5_CTG1_1"                  
#> [136] "CHR_HSCHR18_ALT21_CTG2_1"              
#> [137] "CHR_HSCHR18_ALT2_CTG2_1"               
#> [138] "CHR_HSCHR19KIR_ABC08_A1_HAP_CTG3_1"    
#> [139] "CHR_HSCHR19KIR_ABC08_AB_HAP_C_P_CTG3_1"
#> [140] "CHR_HSCHR19KIR_ABC08_AB_HAP_T_P_CTG3_1"
#> [141] "CHR_HSCHR19KIR_FH05_A_HAP_CTG3_1"      
#> [142] "CHR_HSCHR19KIR_FH05_B_HAP_CTG3_1"      
#> [143] "CHR_HSCHR19KIR_FH06_A_HAP_CTG3_1"      
#> [144] "CHR_HSCHR19KIR_FH06_BA1_HAP_CTG3_1"    
#> [145] "CHR_HSCHR19KIR_FH08_A_HAP_CTG3_1"      
#> [146] "CHR_HSCHR19KIR_FH08_BAX_HAP_CTG3_1"    
#> [147] "CHR_HSCHR19KIR_FH13_A_HAP_CTG3_1"      
#> [148] "CHR_HSCHR19KIR_FH13_BA2_HAP_CTG3_1"    
#> [149] "CHR_HSCHR19KIR_FH15_A_HAP_CTG3_1"      
#> [150] "CHR_HSCHR19KIR_FH15_B_HAP_CTG3_1"      
#> [151] "CHR_HSCHR19KIR_G085_A_HAP_CTG3_1"      
#> [152] "CHR_HSCHR19KIR_G085_BA1_HAP_CTG3_1"    
#> [153] "CHR_HSCHR19KIR_G248_A_HAP_CTG3_1"      
#> [154] "CHR_HSCHR19KIR_G248_BA2_HAP_CTG3_1"    
#> [155] "CHR_HSCHR19KIR_GRC212_AB_HAP_CTG3_1"   
#> [156] "CHR_HSCHR19KIR_GRC212_BA1_HAP_CTG3_1"  
#> [157] "CHR_HSCHR19KIR_LUCE_A_HAP_CTG3_1"      
#> [158] "CHR_HSCHR19KIR_LUCE_BDEL_HAP_CTG3_1"   
#> [159] "CHR_HSCHR19KIR_RP5_B_HAP_CTG3_1"       
#> [160] "CHR_HSCHR19KIR_RSH_A_HAP_CTG3_1"       
#> [161] "CHR_HSCHR19KIR_RSH_BA2_HAP_CTG3_1"     
#> [162] "CHR_HSCHR19KIR_T7526_A_HAP_CTG3_1"     
#> [163] "CHR_HSCHR19KIR_T7526_BDEL_HAP_CTG3_1"  
#> [164] "CHR_HSCHR19LRC_COX1_CTG3_1"            
#> [165] "CHR_HSCHR19LRC_COX2_CTG3_1"            
#> [166] "CHR_HSCHR19LRC_LRC_I_CTG3_1"           
#> [167] "CHR_HSCHR19LRC_LRC_J_CTG3_1"           
#> [168] "CHR_HSCHR19LRC_LRC_S_CTG3_1"           
#> [169] "CHR_HSCHR19LRC_LRC_T_CTG3_1"           
#> [170] "CHR_HSCHR19LRC_PGF1_CTG3_1"            
#> [171] "CHR_HSCHR19LRC_PGF2_CTG3_1"            
#> [172] "CHR_HSCHR19_1_CTG2"                    
#> [173] "CHR_HSCHR19_1_CTG3_1"                  
#> [174] "CHR_HSCHR19_2_CTG2"                    
#> [175] "CHR_HSCHR19_2_CTG3_1"                  
#> [176] "CHR_HSCHR19_3_CTG2"                    
#> [177] "CHR_HSCHR19_3_CTG3_1"                  
#> [178] "CHR_HSCHR19_4_CTG2"                    
#> [179] "CHR_HSCHR19_4_CTG3_1"                  
#> [180] "CHR_HSCHR19_5_CTG2"                    
#> [181] "CHR_HSCHR1_1_CTG11"                    
#> [182] "CHR_HSCHR1_1_CTG3"                     
#> [183] "CHR_HSCHR1_1_CTG31"                    
#> [184] "CHR_HSCHR1_1_CTG32_1"                  
#> [185] "CHR_HSCHR1_2_CTG3"                     
#> [186] "CHR_HSCHR1_2_CTG31"                    
#> [187] "CHR_HSCHR1_2_CTG32_1"                  
#> [188] "CHR_HSCHR1_3_CTG3"                     
#> [189] "CHR_HSCHR1_3_CTG31"                    
#> [190] "CHR_HSCHR1_3_CTG32_1"                  
#> [191] "CHR_HSCHR1_4_CTG3"                     
#> [192] "CHR_HSCHR1_4_CTG31"                    
#> [193] "CHR_HSCHR1_5_CTG3"                     
#> [194] "CHR_HSCHR1_5_CTG32_1"                  
#> [195] "CHR_HSCHR1_ALT2_1_CTG32_1"             
#> [196] "CHR_HSCHR20_1_CTG1"                    
#> [197] "CHR_HSCHR20_1_CTG2"                    
#> [198] "CHR_HSCHR20_1_CTG3"                    
#> [199] "CHR_HSCHR20_1_CTG4"                    
#> [200] "CHR_HSCHR21_2_CTG1_1"                  
#> [201] "CHR_HSCHR21_3_CTG1_1"                  
#> [202] "CHR_HSCHR21_4_CTG1_1"                  
#> [203] "CHR_HSCHR21_5_CTG2"                    
#> [204] "CHR_HSCHR21_6_CTG1_1"                  
#> [205] "CHR_HSCHR21_8_CTG1_1"                  
#> [206] "CHR_HSCHR22_1_CTG1"                    
#> [207] "CHR_HSCHR22_1_CTG2"                    
#> [208] "CHR_HSCHR22_1_CTG3"                    
#> [209] "CHR_HSCHR22_1_CTG4"                    
#> [210] "CHR_HSCHR22_1_CTG5"                    
#> [211] "CHR_HSCHR22_1_CTG6"                    
#> [212] "CHR_HSCHR22_1_CTG7"                    
#> [213] "CHR_HSCHR22_2_CTG1"                    
#> [214] "CHR_HSCHR22_3_CTG1"                    
#> [215] "CHR_HSCHR22_4_CTG1"                    
#> [216] "CHR_HSCHR22_5_CTG1"                    
#> [217] "CHR_HSCHR22_6_CTG1"                    
#> [218] "CHR_HSCHR22_7_CTG1"                    
#> [219] "CHR_HSCHR22_8_CTG1"                    
#> [220] "CHR_HSCHR2_1_CTG1"                     
#> [221] "CHR_HSCHR2_1_CTG15"                    
#> [222] "CHR_HSCHR2_1_CTG5"                     
#> [223] "CHR_HSCHR2_1_CTG7"                     
#> [224] "CHR_HSCHR2_1_CTG7_2"                   
#> [225] "CHR_HSCHR2_2_CTG1"                     
#> [226] "CHR_HSCHR2_2_CTG15"                    
#> [227] "CHR_HSCHR2_2_CTG7"                     
#> [228] "CHR_HSCHR2_2_CTG7_2"                   
#> [229] "CHR_HSCHR2_3_CTG1"                     
#> [230] "CHR_HSCHR2_3_CTG15"                    
#> [231] "CHR_HSCHR2_3_CTG7_2"                   
#> [232] "CHR_HSCHR2_4_CTG1"                     
#> [233] "CHR_HSCHR2_6_CTG7_2"                   
#> [234] "CHR_HSCHR3_1_CTG1"                     
#> [235] "CHR_HSCHR3_1_CTG2_1"                   
#> [236] "CHR_HSCHR3_1_CTG3"                     
#> [237] "CHR_HSCHR3_2_CTG2_1"                   
#> [238] "CHR_HSCHR3_2_CTG3"                     
#> [239] "CHR_HSCHR3_3_CTG1"                     
#> [240] "CHR_HSCHR3_3_CTG3"                     
#> [241] "CHR_HSCHR3_4_CTG2_1"                   
#> [242] "CHR_HSCHR3_4_CTG3"                     
#> [243] "CHR_HSCHR3_5_CTG2_1"                   
#> [244] "CHR_HSCHR3_5_CTG3"                     
#> [245] "CHR_HSCHR3_6_CTG3"                     
#> [246] "CHR_HSCHR3_7_CTG3"                     
#> [247] "CHR_HSCHR3_8_CTG3"                     
#> [248] "CHR_HSCHR3_9_CTG3"                     
#> [249] "CHR_HSCHR4_11_CTG12"                   
#> [250] "CHR_HSCHR4_1_CTG12"                    
#> [251] "CHR_HSCHR4_1_CTG4"                     
#> [252] "CHR_HSCHR4_1_CTG6"                     
#> [253] "CHR_HSCHR4_1_CTG9"                     
#> [254] "CHR_HSCHR4_2_CTG12"                    
#> [255] "CHR_HSCHR4_2_CTG4"                     
#> [256] "CHR_HSCHR4_3_CTG12"                    
#> [257] "CHR_HSCHR4_4_CTG12"                    
#> [258] "CHR_HSCHR4_5_CTG12"                    
#> [259] "CHR_HSCHR4_6_CTG12"                    
#> [260] "CHR_HSCHR4_7_CTG12"                    
#> [261] "CHR_HSCHR4_8_CTG12"                    
#> [262] "CHR_HSCHR4_9_CTG12"                    
#> [263] "CHR_HSCHR5_1_CTG1"                     
#> [264] "CHR_HSCHR5_1_CTG1_1"                   
#> [265] "CHR_HSCHR5_1_CTG5"                     
#> [266] "CHR_HSCHR5_2_CTG1"                     
#> [267] "CHR_HSCHR5_2_CTG1_1"                   
#> [268] "CHR_HSCHR5_2_CTG5"                     
#> [269] "CHR_HSCHR5_3_CTG1"                     
#> [270] "CHR_HSCHR5_3_CTG5"                     
#> [271] "CHR_HSCHR5_4_CTG1"                     
#> [272] "CHR_HSCHR5_4_CTG1_1"                   
#> [273] "CHR_HSCHR5_5_CTG1"                     
#> [274] "CHR_HSCHR5_6_CTG1"                     
#> [275] "CHR_HSCHR5_7_CTG1"                     
#> [276] "CHR_HSCHR6_1_CTG10"                    
#> [277] "CHR_HSCHR6_1_CTG2"                     
#> [278] "CHR_HSCHR6_1_CTG3"                     
#> [279] "CHR_HSCHR6_1_CTG4"                     
#> [280] "CHR_HSCHR6_1_CTG5"                     
#> [281] "CHR_HSCHR6_1_CTG6"                     
#> [282] "CHR_HSCHR6_1_CTG7"                     
#> [283] "CHR_HSCHR6_1_CTG8"                     
#> [284] "CHR_HSCHR6_1_CTG9"                     
#> [285] "CHR_HSCHR6_8_CTG1"                     
#> [286] "CHR_HSCHR6_MHC_APD_CTG1"               
#> [287] "CHR_HSCHR6_MHC_COX_CTG1"               
#> [288] "CHR_HSCHR6_MHC_DBB_CTG1"               
#> [289] "CHR_HSCHR6_MHC_MANN_CTG1"              
#> [290] "CHR_HSCHR6_MHC_MCF_CTG1"               
#> [291] "CHR_HSCHR6_MHC_QBL_CTG1"               
#> [292] "CHR_HSCHR6_MHC_SSTO_CTG1"              
#> [293] "CHR_HSCHR7_1_CTG1"                     
#> [294] "CHR_HSCHR7_1_CTG4_4"                   
#> [295] "CHR_HSCHR7_1_CTG6"                     
#> [296] "CHR_HSCHR7_1_CTG7"                     
#> [297] "CHR_HSCHR7_2_CTG1"                     
#> [298] "CHR_HSCHR7_2_CTG4_4"                   
#> [299] "CHR_HSCHR7_2_CTG6"                     
#> [300] "CHR_HSCHR7_2_CTG7"                     
#> [301] "CHR_HSCHR7_3_CTG6"                     
#> [302] "CHR_HSCHR8_1_CTG1"                     
#> [303] "CHR_HSCHR8_1_CTG6"                     
#> [304] "CHR_HSCHR8_1_CTG7"                     
#> [305] "CHR_HSCHR8_2_CTG1"                     
#> [306] "CHR_HSCHR8_2_CTG7"                     
#> [307] "CHR_HSCHR8_3_CTG1"                     
#> [308] "CHR_HSCHR8_3_CTG7"                     
#> [309] "CHR_HSCHR8_4_CTG1"                     
#> [310] "CHR_HSCHR8_4_CTG7"                     
#> [311] "CHR_HSCHR8_5_CTG1"                     
#> [312] "CHR_HSCHR8_5_CTG7"                     
#> [313] "CHR_HSCHR8_6_CTG1"                     
#> [314] "CHR_HSCHR8_7_CTG1"                     
#> [315] "CHR_HSCHR8_8_CTG1"                     
#> [316] "CHR_HSCHR8_9_CTG1"                     
#> [317] "CHR_HSCHR9_1_CTG1"                     
#> [318] "CHR_HSCHR9_1_CTG2"                     
#> [319] "CHR_HSCHR9_1_CTG3"                     
#> [320] "CHR_HSCHR9_1_CTG4"                     
#> [321] "CHR_HSCHR9_1_CTG5"                     
#> [322] "CHR_HSCHR9_1_CTG6"                     
#> [323] "CHR_HSCHRX_1_CTG3"                     
#> [324] "CHR_HSCHRX_2_CTG12"                    
#> [325] "CHR_HSCHRX_2_CTG3"                     
#> [326] "GL000009.2"                            
#> [327] "GL000194.1"                            
#> [328] "GL000195.1"                            
#> [329] "GL000205.2"                            
#> [330] "GL000213.1"                            
#> [331] "GL000216.2"                            
#> [332] "GL000218.1"                            
#> [333] "GL000219.1"                            
#> [334] "GL000220.1"                            
#> [335] "GL000225.1"                            
#> [336] "KI270442.1"                            
#> [337] "KI270711.1"                            
#> [338] "KI270713.1"                            
#> [339] "KI270721.1"                            
#> [340] "KI270726.1"                            
#> [341] "KI270727.1"                            
#> [342] "KI270728.1"                            
#> [343] "KI270731.1"                            
#> [344] "KI270733.1"                            
#> [345] "KI270734.1"                            
#> [346] "KI270744.1"                            
#> [347] "KI270750.1"                            
#> [348] "LRG_183"                               
#> [349] "LRG_187"                               
#> [350] "LRG_239"                               
#> [351] "LRG_311"                               
#> [352] "LRG_721"                               
#> [353] "LRG_741"                               
#> [354] "LRG_93"                                
#> [355] "chrM"                                  
#> [356] "chrX"                                  
#> [357] "chrY"                                  

## Change the option ensembldb.seqnameNotFound to return NA in case
## the seqname can not be mapped form Ensembl to UCSC.
options(ensembldb.seqnameNotFound = NA)

seqlevels(edb)
#> Warning: More than 5 seqnames with seqlevels style of the database (Ensembl) could not be mapped to the seqlevels style: UCSC!) Returning NA for these.
#>   [1] "chr1"  "chr10" "chr11" "chr12" "chr13" "chr14" "chr15" "chr16" "chr17"
#>  [10] "chr18" "chr19" "chr2"  "chr20" "chr21" "chr22" "chr3"  "chr4"  "chr5" 
#>  [19] "chr6"  "chr7"  "chr8"  "chr9"  NA      NA      NA      NA      NA     
#>  [28] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#>  [37] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#>  [46] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#>  [55] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#>  [64] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#>  [73] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#>  [82] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#>  [91] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [100] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [109] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [118] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [127] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [136] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [145] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [154] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [163] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [172] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [181] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [190] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [199] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [208] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [217] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [226] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [235] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [244] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [253] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [262] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [271] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [280] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [289] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [298] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [307] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [316] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [325] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [334] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [343] NA      NA      NA      NA      NA      NA      NA      NA      NA     
#> [352] NA      NA      NA      "chrM"  "chrX"  "chrY" 

## Restoring the original setting.
options(ensembldb.seqnameNotFound = "ORIGINAL")