List EnsDb databases in a MariaDB/MySQL server

The listEnsDbs function lists EnsDb databases in a MariaDB/MySQL server.

listEnsDbs(dbcon, host, port, user, pass)

Arguments

dbcon
A DBIConnection object providing access to a MariaDB/MySQL database. Either dbcon or all of the other arguments have to be specified.

host
Character specifying the host on which the MySQL server is running.

port
The port of the MariaDB/MySQL server (usually 3306).

user
The username for the MariaDB/MySQL server.

pass
The password for the MariaDB/MySQL server.

Value

A data.frame listing the database names, organism name and Ensembl version of the EnsDb databases found on the server.

Details

The use of this function requires the RMariaDB package to be installed. In addition user credentials to access a MySQL server (with already installed EnsDb databases), or with write access are required. For the latter EnsDb databases can be added with the useMySQL method. EnsDb databases in a MariaDB/MySQL server follow the same naming conventions than EnsDb packages, with the exception that the name is all lower case and that each "." is replaced by "_".

See also

Author

Johannes Rainer

Examples

if (FALSE) {
library(RMariaDB)
dbcon <- dbConnect(MariaDB(), host = "localhost", user = my_user, pass = my_pass)
listEnsDbs(dbcon)
}